add some confetti animation to celebration reaction
